requirements.txt
agora_token_builder==1.0.0
asgiref==3.11.0
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
certifi==2026.1.4
cffi==2.0.0
channels==4.0.0
charset-normalizer==3.4.4
constantly==23.10.4
crispy-bootstrap4==2022.1
cryptography==46.0.4
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.57.0
django-crispy-forms==2.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
oauthlib==3.3.1
packaging==26.0
Pillow==10.0.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
sqlparse==0.5.5
tomli==2.4.0
Twisted==25.5.0
txaio==25.9.2
urllib3==2.6.3
whitenoise==6.6.0
zope.interface==8.2

# ===============================
# ML Dependencies for Sign Language Recognition
# ===============================
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.1.78
mediapipe==0.10.9
protobuf==3.20.3
pandas==2.0.3
absl-py>=1.4.0
flatbuffers>=2.0
sounddevice>=0.4.6

# typing_extensions - Compatible with both TensorFlow and other packages
typing_extensions>=4.2.0,<4.6.0
